challenge 1 complete

+++ capstone/tenmo-server/src/main/java/com/techelevator/tenmo/dao/AccountDao.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface AccountDao {
 
     boolean create(int user_id);
-    Account getAccountsById(int user_id);
+    Account getAccountsById(long user_id);
     void update(int account_id, double balance);
     Account getAccountByAccountId(int account_id);
 }
